feat: CommonBtn type 분리
