+++ components/commonBtn/commonBtn.types.ts
@@ -1,0 +1,18 @@
+/**
+ * 버튼 타입 enum입니다.
+ * @DEACTIVE 비활성화 버튼
+ * @ACTIVE 활성화 버튼
+ */
+export enum ButtonStyle {
+  DEACTIVE = 'deactive',
+  ACTIVE = 'active',
+}
+
+export type ClickHandler = (e: React.MouseEvent) => void
+
+export interface ICommonBtnProps {
+  type?: 'button' | 'submit'
+  style?: ButtonStyle
+  children: React.ReactNode
+  onClick?: ClickHandler
+}
