--- types/enums/commonBtn.enum.ts
@@ -1,9 +1,0 @@
-/**
- * 버튼 타입 enum입니다.
- * @DEACTIVE 비활성화 버튼
- * @ACTIVE 활성화 버튼
- */
-export enum ButtonStyle {
-  DEACTIVE = 'deactive',
-  ACTIVE = 'active',
-}
